Trying to descend a level to grab data, and then ascend is harder to do. So it is better to just go ahead and keep MainFrame on the same level as Pysat_UI everything else can stay where it is. (maybe) 	deleted:    src/Python PYSAT_GUI/PythonUI/mainframe_.py -> src/Python PYSAT_GUI/pysat_ui.py
